fix somethings

+++ src/interface/stargate.sol
@@ -2,26 +2,25 @@
 // Copyright (C) 2023
 // stargate.sol : copy from https://github.com/stargate-protocol/stargate/blob/main/contracts/interfaces/IStargateRouter.sol
 //
-
 pragma solidity ^0.8.20;
 pragma abicoder v2;
 
 interface IStargateRouter {
   struct lzTxObj {
-    uint256 dstGasForCall;
-    uint256 dstNativeAmount;
+    uint dstGasForCall;
+    uint dstNativeAmount;
     bytes dstNativeAddr;
   }
 
-  function addLiquidity(uint256 _poolId, uint256 _amountLD, address _to) external;
+  function addLiquidity(uint _poolId, uint _amountLD, address _to) external;
 
   function swap(
     uint16 _dstChainId,
-    uint256 _srcPoolId,
-    uint256 _dstPoolId,
+    uint _srcPoolId,
+    uint _dstPoolId,
     address payable _refundAddress,
-    uint256 _amountLD,
-    uint256 _minAmountLD,
+    uint _amountLD,
+    uint _minAmountLD,
     lzTxObj memory _lzTxParams,
     bytes calldata _to,
     bytes calldata _payload
@@ -29,35 +28,33 @@ interface IStargateRouter {
 
   function redeemRemote(
     uint16 _dstChainId,
-    uint256 _srcPoolId,
-    uint256 _dstPoolId,
+    uint _srcPoolId,
+    uint _dstPoolId,
     address payable _refundAddress,
-    uint256 _amountLP,
-    uint256 _minAmountLD,
+    uint _amountLP,
+    uint _minAmountLD,
     bytes calldata _to,
     lzTxObj memory _lzTxParams
   ) external payable;
 
-  function instantRedeemLocal(
-    uint16 _srcPoolId,
-    uint256 _amountLP,
-    address _to
-  ) external returns (uint256);
+  function instantRedeemLocal(uint16 _srcPoolId, uint _amountLP, address _to)
+    external
+    returns (uint);
 
   function redeemLocal(
     uint16 _dstChainId,
-    uint256 _srcPoolId,
-    uint256 _dstPoolId,
+    uint _srcPoolId,
+    uint _dstPoolId,
     address payable _refundAddress,
-    uint256 _amountLP,
+    uint _amountLP,
     bytes calldata _to,
     lzTxObj memory _lzTxParams
   ) external payable;
 
   function sendCredits(
     uint16 _dstChainId,
-    uint256 _srcPoolId,
-    uint256 _dstPoolId,
+    uint _srcPoolId,
+    uint _dstPoolId,
     address payable _refundAddress
   ) external payable;
 
@@ -67,7 +64,7 @@ interface IStargateRouter {
     bytes calldata _toAddress,
     bytes calldata _transferAndCallPayload,
     lzTxObj memory _lzTxParams
-  ) external view returns (uint256, uint256);
+  ) external view returns (uint, uint);
 }
 
 interface IRouterETH {
@@ -75,7 +72,7 @@ interface IRouterETH {
     uint16 _dstChainId, // destination Stargate chainId
     address payable _refundAddress, // refund additional messageFee to this address
     bytes calldata _toAddress, // the receiver of the destination ETH
-    uint256 _amountLD, // the amount, in Local Decimals, to be swapped
-    uint256 _minAmountLD // the minimum amount accepted out on destination
+    uint _amountLD, // the amount, in Local Decimals, to be swapped
+    uint _minAmountLD // the minimum amount accepted out on destination
   ) external payable;
 }
