+++ src/interface/swapNFT.sol
@@ -2,14 +2,12 @@
 // Copyright (C) 2023
 // nft.sol : tswap nft interface
 //
-
 pragma solidity ^0.8.20;
+
 import "@openzeppelin/contracts/token/ERC721/IERC721.sol";
 
 interface ISwapNFT is IERC721 {
-  function positions(
-    uint256 tokenId
-  )
+  function positions(uint tokenId)
     external
     view
     returns (
@@ -21,8 +19,8 @@ interface ISwapNFT is IERC721 {
       int24 tickLower,
       int24 tickUpper,
       uint128 liquidity,
-      uint256 feeGrowthInside0LastX128,
-      uint256 feeGrowthInside1LastX128,
+      uint feeGrowthInside0LastX128,
+      uint feeGrowthInside1LastX128,
       uint128 tokensOwed0,
       uint128 tokensOwed1
     );
@@ -33,53 +31,57 @@ interface ISwapNFT is IERC721 {
     uint24 fee;
     int24 tickLower;
     int24 tickUpper;
-    uint256 amount0Desired;
-    uint256 amount1Desired;
-    uint256 amount0Min;
-    uint256 amount1Min;
+    uint amount0Desired;
+    uint amount1Desired;
+    uint amount0Min;
+    uint amount1Min;
     address recipient;
-    uint256 deadline;
+    uint deadline;
   }
 
-  function mint(
-    MintParams calldata params
-  ) external payable returns (uint256 tokenId, uint128 liquidity, uint256 amount0, uint256 amount1);
+  function mint(MintParams calldata params)
+    external
+    payable
+    returns (uint tokenId, uint128 liquidity, uint amount0, uint amount1);
 
   struct IncreaseLiquidityParams {
-    uint256 tokenId;
-    uint256 amount0Desired;
-    uint256 amount1Desired;
-    uint256 amount0Min;
-    uint256 amount1Min;
-    uint256 deadline;
+    uint tokenId;
+    uint amount0Desired;
+    uint amount1Desired;
+    uint amount0Min;
+    uint amount1Min;
+    uint deadline;
   }
 
-  function increaseLiquidity(
-    IncreaseLiquidityParams calldata params
-  ) external payable returns (uint128 liquidity, uint256 amount0, uint256 amount1);
+  function increaseLiquidity(IncreaseLiquidityParams calldata params)
+    external
+    payable
+    returns (uint128 liquidity, uint amount0, uint amount1);
 
   struct DecreaseLiquidityParams {
-    uint256 tokenId;
+    uint tokenId;
     uint128 liquidity;
-    uint256 amount0Min;
-    uint256 amount1Min;
-    uint256 deadline;
+    uint amount0Min;
+    uint amount1Min;
+    uint deadline;
   }
 
-  function decreaseLiquidity(
-    DecreaseLiquidityParams calldata params
-  ) external payable returns (uint256 amount0, uint256 amount1);
+  function decreaseLiquidity(DecreaseLiquidityParams calldata params)
+    external
+    payable
+    returns (uint amount0, uint amount1);
 
   struct CollectParams {
-    uint256 tokenId;
+    uint tokenId;
     address recipient;
     uint128 amount0Max;
     uint128 amount1Max;
   }
 
-  function collect(
-    CollectParams calldata params
-  ) external payable returns (uint256 amount0, uint256 amount1);
+  function collect(CollectParams calldata params)
+    external
+    payable
+    returns (uint amount0, uint amount1);
 
-  function burn(uint256 tokenId) external payable;
+  function burn(uint tokenId) external payable;
 }
