+++ src/interface/swaprouter.sol
@@ -2,7 +2,6 @@
 // Copyright (C) 2023
 // swaprouter.sol : tswap router interface
 //
-
 pragma solidity ^0.8.20;
 
 interface ISwapRouter {
@@ -11,16 +10,17 @@ interface ISwapRouter {
     address tokenOut;
     uint24 fee;
     address recipient;
-    uint256 deadline;
-    uint256 amountOut;
-    uint256 amountInMaximum;
+    uint deadline;
+    uint amountOut;
+    uint amountInMaximum;
     uint160 sqrtPriceLimitX96;
   }
 
   /// @notice Swaps as little as possible of one token for `amountOut` of another token
   /// @param params The parameters necessary for the swap, encoded as `ExactOutputSingleParams` in calldata
   /// @return amountIn The amount of the input token
-  function exactOutputSingle(
-    ExactOutputSingleParams calldata params
-  ) external payable returns (uint256 amountIn);
+  function exactOutputSingle(ExactOutputSingleParams calldata params)
+    external
+    payable
+    returns (uint amountIn);
 }
